change defaults, new temporary tests
