+++ .github/workflows/fuzzers.yml
@@ -19,8 +19,6 @@ jobs:
           - aflfast
           - aflplusplus
           - aflplusplus_optimal
-          - aflplusplus_shmem
-          - aflplusplus_optimal_shmem
           - aflsmart
           - ankou
           - entropic
@@ -37,6 +35,12 @@ jobs:
           - afl_qemu
           - aflplusplus_qemu
           - honggfuzz_qemu
+          # Temporary variants.
+          - aflplusplus_lto_dict
+          - aflplusplus_lto
+          - aflplusplus_ltoinstrim
+          - aflplusplus_ctx
+          - aflplusplus_ngram2
 
         benchmark_type:
           - oss-fuzz
