+++ fuzzers/aflplusplus_lto/fuzzer.py
@@ -1,0 +1,60 @@
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Integration code for AFLplusplus fuzzer."""
+
+# This optimized afl++ variant should always be run together with
+# "aflplusplus" to show the difference - a default configured afl++ vs.
+# a hand-crafted optimized one. afl++ is configured not to enable the good
+# stuff by default to be as close to vanilla afl as possible.
+# But this means that the good stuff is hidden away in this benchmark
+# otherwise.
+
+import os
+import shutil
+import glob
+
+from fuzzers.aflplusplus import fuzzer as aflplusplus_fuzzer
+
+
+def build():  # pylint: disable=too-many-branches,too-many-statements
+    """Build benchmark."""
+    benchmark_name = os.environ['BENCHMARK']
+
+    if benchmark_name == 'bloaty_fuzz_target':
+        aflplusplus_fuzzer.build('lto', 'dynamic', 'nozero')
+    elif benchmark_name == 'curl_curl_fuzzer_http':
+        aflplusplus_fuzzer.build('lto', 'dynamic', 'nozero')
+    elif benchmark_name == 'freetype2-2017':
+        aflplusplus_fuzzer.build('lto', 'dynamic', 'nozero')
+    elif benchmark_name == 'jsoncpp_jsoncpp_fuzzer':
+        aflplusplus_fuzzer.build('lto', 'dynamic', 'nozero')
+    elif benchmark_name == 'openssl_x509':
+        aflplusplus_fuzzer.build('lto', 'dynamic', 'nozero')
+    elif benchmark_name == 'php_php-fuzz-parser':
+        aflplusplus_fuzzer.build('lto', 'dynamic', 'nozero')
+    else:
+        aflplusplus_fuzzer.build('lto', 'nozero')
+
+    for copy_file in glob.glob("/afl/libc*"):
+        shutil.copy(copy_file, os.environ['OUT'])
+
+
+def fuzz(input_corpus, output_corpus, target_binary):
+    """Run fuzzer."""
+    run_options = []
+
+    aflplusplus_fuzzer.fuzz(input_corpus,
+                            output_corpus,
+                            target_binary,
+                            flags=(run_options))
