+++ fuzzers/aflplusplus_ltoinstrim/builder.Dockerfile
