+++ fuzzers/aflplusplus_ltoinstrim/runner.Dockerfile
