+++ fuzzers/aflplusplus_ngram2/fuzzer.py
@@ -13,38 +13,31 @@
 # limitations under the License.
 """Integration code for AFLplusplus fuzzer."""
 
+# This optimized afl++ variant should always be run together with
+# "aflplusplus" to show the difference - a default configured afl++ vs.
+# a hand-crafted optimized one. afl++ is configured not to enable the good
+# stuff by default to be as close to vanilla afl as possible.
+# But this means that the good stuff is hidden away in this benchmark
+# otherwise.
+
 import os
-import subprocess
+import shutil
+import glob
 
 from fuzzers.aflplusplus import fuzzer as aflplusplus_fuzzer
 
 
 def build():
     """Build benchmark."""
-    aflplusplus_fuzzer.build('qemu')
+
+    aflplusplus_fuzzer.build('classic', 'ngram2', 'nozero')
 
 
 def fuzz(input_corpus, output_corpus, target_binary):
     """Run fuzzer."""
-    # Get LLVMFuzzerTestOneInput address.
-    nm_proc = subprocess.run([
-        'sh', '-c',
-        'nm \'' + target_binary + '\' | grep \'T afl_qemu_driver_stdin_input\''
-    ],
-                             stdout=subprocess.PIPE,
-                             check=True)
-    target_func = "0x" + nm_proc.stdout.split()[0].decode("utf-8")
-    print('[fuzz] afl_qemu_driver_stdin_input() address =', target_func)
+    run_options = []
 
-    # Fuzzer option for qemu_mode.
-    flags = ['-Q'] 
-
-    os.environ['AFL_COMPCOV_LEVEL'] = '3'  # Complete compcov including floats
-    os.environ['AFL_QEMU_PERSISTENT_ADDR'] = target_func
-    os.environ['AFL_ENTRYPOINT'] = target_func
-    os.environ['AFL_QEMU_PERSISTENT_CNT'] = "100000"
-    os.environ['AFL_QEMU_DRIVER_NO_HOOK'] = "1"
     aflplusplus_fuzzer.fuzz(input_corpus,
                             output_corpus,
                             target_binary,
-                            flags=flags)
+                            flags=(run_options))
