+++ fuzzers/aflplusplus_qemu/builder.Dockerfile
@@ -24,7 +24,7 @@ RUN apt-get update && \
 # Build without Python support as we don't need it.
 # Set AFL_NO_X86 to skip flaky tests.
 RUN git clone https://github.com/AFLplusplus/AFLplusplus.git /afl && \
-    cd /afl && git checkout e5e485fcdb039fc77842b0753a4adf42d6063388 && \
+    cd /afl && git checkout 0aed549df102cde6a60dc9ef57524413e978814f && \
     unset CFLAGS && unset CXXFLAGS && \
     AFL_NO_X86=1 CC=clang PYTHON_INCLUDE=/ make && \
     cd qemu_mode && ./build_qemu_support.sh && cd .. && \
